+++ src/main/java/hello/controllers/HomeController.java
@@ -1,4 +1,4 @@
-package hello;
+package hello.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
@@ -16,7 +16,7 @@ import hello.geojson.FeatureCollection;
 import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
 
 @Controller
-public class WebController {
+public class HomeController {
 
     @Autowired
     private ClientRegistrationRepository clientRegistrationRepository;
@@ -40,32 +40,4 @@ public class WebController {
         return "login";
     }
 
-    @GetMapping("/page1")
-    public String getPage1(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-
-        return "page1";
-    }
-
-    @GetMapping("/page2")
-    public String getPage2(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-
-        return "page2";
-    }
-
-    @GetMapping("/earthquakes/search")
-    public String getEarthquakesSearch(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken, EqSearch eqSearch) {
-        return "earthquakes/search";
-    }
-
-    @GetMapping("/earthquakes/results")
-    public String getEarthquakesResults(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken, EqSearch eqSearch) {
-        EarthquakeQueryService e = new EarthquakeQueryService();
-        
-        model.addAttribute("eqSearch", eqSearch);
-        String json = e.getJSON(eqSearch.getDistance(), eqSearch.getMinmag());
-        model.addAttribute("json", json);
-        FeatureCollection featureCollection = FeatureCollection.fromJSON(json);
-        model.addAttribute("featureCollection",featureCollection);
-        return "earthquakes/results";
-    }
 }
